Fix configstore on Pixel 1
Old configstore seccomp policy are broken by new bionic, calling gettid.
We could modify libminijail to explicitly include gettid... except
minijail is in vendor.
We can't replace minijail with ours because it is in C++ so no stable
ABI.
Anyway, configstore is perfectly safe, if it has a security flaw, every
other service has it.
So we can use an escape hatch provided in AOSP configstore source code:
if the file is unreadable don't call the jail at all.